rename existing grid and create new `Grid` class and spec
